+++ styleguide.config.js
@@ -1,6 +1,7 @@
 /* eslint-env node */
 
 module.exports = {
+    applyBabelToTypescriptCode: true,
     getExceptionForLoaders() {
         return {
             jsLoader: [/spec\.jsx$/],